+++ ooniprobe/Utility/DictionaryUtility.m
@@ -1,0 +1,42 @@
+// Part of MeasurementKit <https://measurement-kit.github.io/>.
+// MeasurementKit is free software. See AUTHORS and LICENSE for more
+// information on the copying conditions.
+
+#import "DictionaryUtility.h"
+
+@implementation DictionaryUtility
+
+
++ (NSDictionary *)parseQueryString:(NSString *)query {
+    NSMutableDictionary *dict = [[NSMutableDictionary alloc] init];
+    NSArray *pairs = [query componentsSeparatedByString:@"&"];
+    
+    for (NSString *pair in pairs) {
+        NSArray *elements = [pair componentsSeparatedByString:@"="];
+        NSString *key = [[elements objectAtIndex:0] stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
+        NSString *val = [[elements objectAtIndex:1] stringByReplacingPercentEscapesUsingEncoding:NSUTF8StringEncoding];
+        
+        [dict setObject:val forKey:key];
+    }
+    return dict;
+}
+
++ (NSDictionary*)getParametersFromDict:(NSDictionary*)dict{
+    NSMutableDictionary *parameters = [[NSMutableDictionary alloc] init];
+    for (NSString *key in [dict allKeys]){
+        NSString *current = [dict objectForKey:key];
+        NSError *error;
+        NSData *data = [current dataUsingEncoding:NSUTF8StringEncoding];
+        NSDictionary *dictionary = [NSJSONSerialization JSONObjectWithData:data options:kNilOptions error:&error];
+        if (error == nil){
+            //set a dictionary subparameter
+            [parameters setObject:dictionary forKey:key];
+        }
+        else{
+            //set a string subparameter
+            [parameters setObject:current forKey:key];
+        }
+    }
+    return parameters;
+}
+@end