+++ ooniprobe/View/BrowserViewController.h
@@ -1,10 +1,6 @@
-//
-//  BrowserViewController.h
-//  ooniprobe
-//
-//  Created by Lorenzo Primiterra on 23/06/17.
-//  Copyright © 2017 Simone Basso. All rights reserved.
-//
+// Part of MeasurementKit <https://measurement-kit.github.io/>.
+// MeasurementKit is free software. See AUTHORS and LICENSE for more
+// information on the copying conditions.
 
 #import <UIKit/UIKit.h>