+++ ooniprobe/View/SettingsTableViewController.mm
@@ -3,6 +3,7 @@
 // information on the copying conditions.
 
 #import "SettingsTableViewController.h"
+#define alert_tag_collector_address 1
 
 @interface SettingsTableViewController ()
 @property (readwrite) IBOutlet UIBarButtonItem* revealButtonItem;
@@ -81,6 +82,12 @@
     return NSLocalizedString(@"advanced_settings", nil);
 }
 
+-(void)tableView:(UITableView *)tableView willDisplayHeaderView:(UIView *)view forSection:(NSInteger)section
+{
+    UITableViewHeaderFooterView *header = (UITableViewHeaderFooterView *)view;
+    header.textLabel.font = [UIFont fontWithName:@"FiraSansOT-Bold" size:15];
+    header.textLabel.textColor = color_off_black;
+}
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
     UITableViewCell *cell;
@@ -224,8 +231,8 @@
 {
     if (indexPath.section == 2 && indexPath.row == [advancedItems count] -1){
         NSString *current = [advancedItems objectAtIndex:indexPath.row];
-        UIAlertView *alert = [[UIAlertView alloc] initWithTitle:NSLocalizedString(current, @"") message:nil delegate:self cancelButtonTitle:NSLocalizedString(@"cancel", @"") otherButtonTitles:NSLocalizedString(@"ok", nil), nil];
-        alert.tag = indexPath.row;
+        UIAlertView *alert = [[UIAlertView alloc] initWithTitle:NSLocalizedString(current, @"") message:nil delegate:self cancelButtonTitle:NSLocalizedString(@"cancel", @"") otherButtonTitles:NSLocalizedString(@"ok", nil), NSLocalizedString(@"set_default", nil), nil];
+        alert.tag = alert_tag_collector_address;
         alert.alertViewStyle = UIAlertViewStylePlainTextInput;
         value = [alert textFieldAtIndex:0];
         value.text = [[NSUserDefaults standardUserDefaults] objectForKey:@"collector_address"];
@@ -239,11 +246,18 @@
 
 - (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex
 {
-    if (buttonIndex == 1 && value.text.length > 0) {
-        NSString *current = [advancedItems objectAtIndex:alertView.tag];
-        [[NSUserDefaults standardUserDefaults] setObject:value.text forKey:current];
-        [[NSUserDefaults standardUserDefaults] synchronize];
-        [self.tableView reloadData];
+    if (alertView.tag == alert_tag_collector_address){
+        //collector address
+        if (buttonIndex == 1 && value.text.length > 0) {
+            [[NSUserDefaults standardUserDefaults] setObject:value.text forKey:@"collector_address"];
+            [[NSUserDefaults standardUserDefaults] synchronize];
+            [self.tableView reloadData];
+        }
+        else if (buttonIndex == 2){
+            [[NSUserDefaults standardUserDefaults] setObject:COLLECTOR_ADDRESS forKey:@"collector_address"];
+            [[NSUserDefaults standardUserDefaults] synchronize];
+            [self.tableView reloadData];
+        }
     }
 }